Updated Rocky and removed Terraform
Updated rocky to use the latest version and am using the long version string so I can trigger a build when a new version of the container comes out.
Removed terraform as we are no longer using it. I go back and forth so it might appear back in a future version.